requirements.txt
GPUtil==1.4.0
pandas~=1.4.2
Pillow~=9.1.0
tensorflow-gpu~=2.8.0
numpy~=1.22.3
opencv-python~=4.5.5.64
matplotlib~=3.5.2
mesh-to-sdf==0.0.15
scikit-learn~=1.0.2
scipy~=1.4.1
spektral~=1.1.0
scikit-video~=1.1.11
trimesh~=4.10.1
vtk~=9.5.2
ipython~=8.3.0
chumpy~=0.70
imageio~=2.19.1
networkx~=2.8
tqdm~=4.64.0
xvfbwrapper~=0.2.9
